+++ lightning_pose/api/model.py
@@ -17,6 +17,7 @@ from lightning_pose.data import _IMAGENET_MEAN, _IMAGENET_STD
 from lightning_pose.data.datamodules import BaseDataModule, UnlabeledDataModule
 from lightning_pose.data.datatypes import MultiviewPredictionResult, PredictionResult
 from lightning_pose.data.utils import convert_bbox_coords
+from lightning_pose.metrics import compute_metrics_single
 from lightning_pose.models import ALLOWED_MODELS
 from lightning_pose.utils import io as io_utils
 from lightning_pose.utils.predictions import generate_labeled_video as generate_labeled_video_fn
@@ -26,7 +27,6 @@ from lightning_pose.utils.predictions import (
     predict_video,
 )
 from lightning_pose.utils.scripts import (
-    compute_metrics_single,
     get_data_module,
     get_dataset,
     get_imgaug_transform,
