+++ lightning_pose/utils/scripts.py
@@ -8,7 +8,6 @@ from pathlib import Path
 
 import imgaug.augmenters as iaa
 import numpy as np
-import pandas as pd
 import torch
 from lightning.pytorch.callbacks import EarlyStopping, LearningRateMonitor, ModelCheckpoint
 from omegaconf import DictConfig, ListConfig, OmegaConf
@@ -30,21 +29,13 @@ from lightning_pose.data.datasets import (
     HeatmapDataset,
     MultiviewHeatmapDataset,
 )
-from lightning_pose.data.datatypes import ComputeMetricsSingleResult
 from lightning_pose.losses.factory import LossFactory
-from lightning_pose.metrics import (
-    pca_multiview_reprojection_error,
-    pca_singleview_reprojection_error,
-    pixel_error,
-    temporal_norm,
-)
 from lightning_pose.models import ALLOWED_MODELS
 from lightning_pose.models.base import (
     _apply_defaults_for_lr_scheduler_params,
     _apply_defaults_for_optimizer_params,
 )
 from lightning_pose.utils import io as io_utils
-from lightning_pose.utils.pca import KeypointPCA
 
 # to ignore imports for sphix-autoapidoc
 __all__ = [
@@ -723,154 +714,3 @@ def calculate_steps_per_epoch(data_module: BaseDataModule) -> int:
     if is_unsupervised:
         steps_per_epoch = max(10, steps_per_epoch)
     return steps_per_epoch
-
-
-def compute_metrics_single(
-    cfg: DictConfig | ListConfig,
-    labels_file: str | Path | None,
-    preds_file: str | Path,
-    data_module: BaseDataModule | UnlabeledDataModule | None = None,
-) -> ComputeMetricsSingleResult:
-    """Compute various metrics on a predictions csv file from a single view."""
-    # load predictions
-    pred_df = pd.read_csv(preds_file, header=[0, 1, 2], index_col=0)
-    keypoint_names = io_utils.get_keypoint_names(
-        cfg, csv_file=str(preds_file), header_rows=[0, 1, 2])
-    xyl_mask = pred_df.columns.get_level_values("coords").isin(["x", "y", "likelihood"])
-    tmp = pred_df.loc[:, xyl_mask].to_numpy().reshape(pred_df.shape[0], -1, 3)
-
-    index = pred_df.index
-    if pred_df.keys()[-1][0] == "set":
-        # these are predictions on labeled data
-        # get rid of last column that contains info about train/val/test set
-        is_video = False
-        set = pred_df.iloc[:, -1].to_numpy()
-    else:
-        # these are predictions on video data
-        is_video = True
-        set = None
-
-    keypoints_pred = tmp[:, :, :2]  # shape (samples, n_keypoints, 2)
-    # confidences = tmp[:, :, -1]  # shape (samples, n_keypoints)
-
-    # hard-code metrics for now
-    if is_video:
-        metrics_to_compute = ["temporal"]
-    else:  # labeled data
-        assert labels_file is not None
-        metrics_to_compute = ["pixel_error"]
-    # for either labeled and unlabeled data, if a pca loss is specified in config, we compute the
-    # associated metric
-
-    if (
-        data_module is not None
-        and cfg.data.get("columns_for_singleview_pca", None) is not None
-        and len(cfg.data.columns_for_singleview_pca) != 0
-        and not isinstance(data_module.dataset, MultiviewHeatmapDataset)  # mirrored-only for now
-    ):
-        metrics_to_compute += ["pca_singleview"]
-    if (
-        data_module is not None
-        and cfg.data.get("mirrored_column_matches", None) is not None
-        and len(cfg.data.mirrored_column_matches) != 0
-        and not isinstance(data_module.dataset, MultiviewHeatmapDataset)  # mirrored-only for now
-    ):
-        metrics_to_compute += ["pca_multiview"]
-
-    result = ComputeMetricsSingleResult()
-    preds_file_path = Path(preds_file)
-    # compute metrics; csv files will be saved to the same directory the prdictions are stored in
-    if "pixel_error" in metrics_to_compute:
-        # Read labeled data
-        assert labels_file is not None, '"pixel_error" metric requires labels_file'
-        labels_df = pd.read_csv(labels_file, header=[0, 1, 2], index_col=0)
-        labels_df = io_utils.fix_empty_first_row(labels_df)
-        assert labels_df.index.equals(index)
-
-        keypoints_true = labels_df.to_numpy().reshape(labels_df.shape[0], -1, 2)
-        error_per_keypoint = pixel_error(keypoints_true, keypoints_pred)
-        error_df = pd.DataFrame(
-            error_per_keypoint, index=pd.Index(index), columns=pd.Index(keypoint_names)
-        )
-        # add train/val/test split
-        if set is not None:
-            error_df["set"] = set
-
-        save_file = preds_file_path.with_name(preds_file_path.stem + "_pixel_error.csv")
-        error_df.to_csv(save_file)
-        result.pixel_error_df = error_df
-
-    if "temporal" in metrics_to_compute:
-        temporal_norm_per_keypoint = temporal_norm(keypoints_pred)
-        temporal_norm_df = pd.DataFrame(
-            temporal_norm_per_keypoint, index=pd.Index(index), columns=pd.Index(keypoint_names)
-        )
-        # add train/val/test split
-        if set is not None:
-            temporal_norm_df["set"] = set
-        save_file = preds_file_path.with_name(preds_file_path.stem + "_temporal_norm.csv")
-        temporal_norm_df.to_csv(save_file)
-        result.temporal_norm_df = temporal_norm_df
-
-    if "pca_singleview" in metrics_to_compute:
-        try:
-            # build pca object
-            assert data_module is not None
-            pca = KeypointPCA(
-                loss_type="pca_singleview",
-                data_module=data_module,
-                components_to_keep=cfg.losses.pca_singleview.components_to_keep,
-                empirical_epsilon_percentile=cfg.losses.pca_singleview.get(
-                    "empirical_epsilon_percentile", 1.0),
-                columns_for_singleview_pca=cfg.data.columns_for_singleview_pca,
-                centering_method=cfg.losses.pca_singleview.get("centering_method", None),
-            )
-            # re-fit pca on the labeled data to get params
-            pca()
-            # compute reprojection error
-            pcasv_error_per_keypoint = pca_singleview_reprojection_error(keypoints_pred, pca)
-            pcasv_df = pd.DataFrame(
-                pcasv_error_per_keypoint, index=pd.Index(index), columns=pd.Index(keypoint_names)
-            )
-            # add train/val/test split
-            if set is not None:
-                pcasv_df["set"] = set
-            save_file = preds_file_path.with_name(
-                preds_file_path.stem + "_pca_singleview_error.csv"
-            )
-            pcasv_df.to_csv(save_file)
-            result.pca_sv_df = pcasv_df
-
-        except ValueError as e:
-            # PCA will fail if not enough train frames.
-            # skip pca metric in this case.
-            # re-raise if this is not the PCA error this try is intended to swallow
-            if "cannot fit PCA" not in str(e):
-                raise e
-
-    if "pca_multiview" in metrics_to_compute:
-        # build pca object
-        assert data_module is not None
-        pca = KeypointPCA(
-            loss_type="pca_multiview",
-            data_module=data_module,
-            components_to_keep=cfg.losses.pca_singleview.components_to_keep,
-            empirical_epsilon_percentile=cfg.losses.pca_singleview.get(
-                "empirical_epsilon_percentile", 1.0),
-            mirrored_column_matches=cfg.data.mirrored_column_matches,
-        )
-        # re-fit pca on the labeled data to get params
-        pca()
-        # compute reprojection error
-        pcamv_error_per_keypoint = pca_multiview_reprojection_error(keypoints_pred, pca)
-        pcamv_df = pd.DataFrame(
-            pcamv_error_per_keypoint, index=pd.Index(index), columns=pd.Index(keypoint_names)
-        )
-        # add train/val/test split
-        if set is not None:
-            pcamv_df["set"] = set
-        save_file = preds_file_path.with_name(preds_file_path.stem + "_pca_multiview_error.csv")
-        pcamv_df.to_csv(save_file)
-        result.pca_mv_df = pcamv_df
-
-    return result
